requirements.txt
requests
python-dotenv
selenium
beautifulsoup4 
google.genai
python-dotenv 